+++ app/Http/Controllers/LoginController.php
@@ -39,7 +39,7 @@ class LoginController extends Controller
     {
         $request->validate([
             // 'name' => ['required', 'string', 'max:255'],
-            'email' => ['required', 'string', 'email', 'max:255', 'unique:' . User::class],
+            'email' => ['required', 'string', 'email', 'max:255', 'unique:'.User::class],
             // 'password' => ['required', Rules\Password::defaults()],
             // 'password' => ['required', 'confirmed', Rules\Password::defaults()],
         ]);
